--- packages/apps/kadena-docs/src/components/Markdown/UnorderdList/index.ts
@@ -1,1 +1,0 @@
-export { UnorderdList } from './UnorderdList';
