+++ packages/apps/kadena-docs/src/components/Markdown/UnorderedList/index.ts
@@ -1,1 +1,1 @@
-export { UnorderedList } from './UnorderdList';
+export { UnorderedList } from './UnorderedList';
